app/requirements.txt
cryptography
Jinja2
PyJWT
requests
Flask
waitress